updated robot code

+++ drivetrain/Core/Inc/PID.h
@@ -49,7 +49,7 @@ typedef struct _PID_TypeDef
 	uint8_t dtime;
 }PID_TypeDef;
 
-	void pid_init(PID_TypeDef* pid, uint16_t maxout, uint16_t integral_limit, float deadband, int16_t  target, float kp, float ki, float kd);
+	void pid_init(PID_TypeDef* pid, float maxout, float integral_limit, float deadband, float  target, float kp, float ki, float kd);
 	void pid_set_constants(PID_TypeDef *pid, float kp,float ki, float kd);
 	float pid_calculate (PID_TypeDef *pid, float measure);
 #endif
